+++ blocks/event-info-component/event-info-component.css
@@ -1,0 +1,58 @@
+.event-info-component input[type="text"],
+.event-info-component textarea {
+  border: none;
+  border-bottom: 1px solid var(--color-black);
+  margin-bottom: 24px;
+  font-family: var(--body-font-family);
+  font-size: var(--type-body-m-size);;
+  width: 100%;
+}
+
+.event-info-component input[type="text"] {
+  font-size: var(--type-heading-l-size);
+  padding: 24px 0;
+}
+
+.event-info-component textarea {
+  resize: vertical;
+  min-height: 40px;
+}
+
+.event-info-component label {
+  font-weight: 700;
+}
+
+.event-info-component .date-time-row {
+  display: flex;
+  align-items: center;
+  gap: 24px;
+  margin-bottom: 24px;
+}
+
+.event-info-component .date-time-row > .date-picker {
+  border-bottom: 1px solid var(--color-black);
+}
+
+.event-info-component .date-time-row > div:last-of-type {
+  flex-grow: 1;
+}
+
+.event-info-component .date-time-row input {
+  border: none;
+  margin-left: 24px;
+}
+
+.event-info-component .time-pickers {
+  display: flex;
+  gap: 16px;
+}
+
+.event-info-component .time-pickers .time-picker-wrapper {
+  flex: 1 1 0;
+  display: flex;
+  gap: 8px;
+}
+
+.event-info-component .time-pickers .time-picker-wrapper select {
+  flex-grow: 1;
+}
