+++ blocks/event-info-component/event-info-component.js
@@ -1,26 +1,7 @@
 import { getLibs } from '../../scripts/utils.js';
-import { handlize } from '../../utils/utils.js';
+import { handlize, standardizeFormComponentHeading } from '../../utils/utils.js';
 
 const { createTag } = await import(`${getLibs()}/utils/utils.js`);
-
-function decorateHeading(el) {
-  const h2 = el.querySelector(':scope > div:first-of-type h2');
-
-  if (h2) {
-    const em = el.querySelector('p > em');
-
-    if (em) {
-      const tooltipText = em.textContent.trim();
-      const toolTipIcon = createTag('span', { class: 'event-heading-tooltip-icon' }, 'i');
-      const toolTipBox = createTag('div', { class: 'event-heading-tooltip-box' }, tooltipText);
-      const toolTipWrapper = createTag('div', { class: 'event-heading-tooltip-wrapper' });
-
-      toolTipWrapper.append(toolTipIcon, toolTipBox);
-      h2.parentElement?.append(toolTipWrapper);
-      em.parentElement?.remove();
-    }
-  }
-}
 
 function decorateTextFields(row) {
   row.classList.add('text=field-row');
@@ -110,7 +91,7 @@ function decorateDateTimeFields(row) {
 
 export default function init(el) {
   el.classList.add('form-component');
-  decorateHeading(el);
+  standardizeFormComponentHeading(el);
 
   const rows = el.querySelectorAll(':scope > div');
   rows.forEach((r, i) => {
