--- blocks/event-information-component/event-information-component.css
@@ -1,62 +1,0 @@
-.event-information-component input[type="text"],
-.event-information-component textarea {
-  border: none;
-  border-bottom: 1px solid var(--color-black);
-  margin-bottom: 24px;
-  font-family: var(--body-font-family);
-  font-size: var(--type-body-m-size);;
-  width: 100%;
-}
-
-.event-information-component input[type="text"] {
-  font-size: var(--type-heading-l-size);
-  padding: 24px 0;
-}
-
-.event-information-component input[type="date"]:focus::-webkit-datetime-edit {
-  color: rgb(255 255 255 / 46%); 
- }
-
-.event-information-component textarea {
-  resize: vertical;
-  min-height: 40px;
-}
-
-.event-information-component label {
-  font-weight: 700;
-}
-
-.event-information-component .date-time-row {
-  display: flex;
-  align-items: center;
-  gap: 24px;
-  margin-bottom: 24px;
-}
-
-.event-information-component .date-time-row > .date-picker {
-  border-bottom: 1px solid var(--color-black);
-}
-
-.event-information-component .date-time-row > div:last-of-type {
-  flex-grow: 1;
-}
-
-.event-information-component .date-time-row input {
-  border: none;
-  margin-left: 24px;
-}
-
-.event-information-component .time-pickers {
-  display: flex;
-  gap: 16px;
-}
-
-.event-information-component .time-pickers .time-picker-wrapper {
-  flex: 1 1 0;
-  display: flex;
-  gap: 8px;
-}
-
-.event-information-component .time-pickers .time-picker-wrapper select {
-  flex-grow: 1;
-}
