+++ blocks/form-handler/form-handler.css
@@ -12,6 +12,7 @@
   flex-grow: 1;
 }
 
+/* TODO: make cards based on card-? class to restructure the form */
 .form-handler .main-frame .section {
   padding: 24px;
   border-radius: 24px;
@@ -24,8 +25,8 @@
 }
 
 .form-handler .event-heading-tooltip-wrapper .event-heading-tooltip-icon {
-  height: 22px;
-  width: 22px;
+  height: 16px;
+  width: 16px;
   display: flex;
   align-items: center;
   justify-content: center;
@@ -35,10 +36,16 @@
   cursor: help;
 }
 
-.form-handler .form-component > div:first-of-type > div {
+.form-handler .form-component > div:first-of-type > div > h2 {
   display: flex;
   align-items: center;
   gap: 8px;
+  margin: 24px 0 0;
+}
+
+.form-handler .form-component > div:first-of-type > div > h2 * {
+  font-size: var(--type-body-s-size);
+  font-weight: normal;
 }
 
 .form-handler .event-heading-tooltip-wrapper {
