+++ scripts/utils.js
@@ -20,7 +20,7 @@
  * Note: This file should have no self-invoking functions.
  * ------------------------------------------------------------
  */
-import { autoUpdateContent } from '../utils/utils.js';
+import autoUpdateContent from '../utils/content-update.js';
 
 export const [setLibs, getLibs] = (() => {
   let libs;
